requirements.txt
# Orchestrator (local box) dependencies
# Worker runtime has its own requirements.txt inside worker_runtime/

pydantic>=2.0
fastapi>=0.110
uvicorn>=0.27
httpx>=0.27
redis>=5.0
boto3>=1.34
pyyaml>=6.0

# Reranker (CPU-friendly ML)
torch>=2.1
open-clip-torch>=2.24
opencv-python>=4.8
numpy>=1.24
scikit-image>=0.22

# Planner
anthropic>=0.39

# Utility
python-dotenv>=1.0
rich>=13.7
